refactored the spotify code
